requirements.txt
pyperclip==1.8.2
certifi==2022.6.15
pytest==7.1.2
PyYAML==6.0
requests==2.28.1
selenium==4.2.0
msgpack==1.0.4
Pillow==9.1.1
lxml==4.9.0
beautifulsoup4==4.11.1
python-gitlab==3.8.1
allure-pytest==2.9.45
pytest-xdist==2.5.0
xattr==0.9.9
ImageHash==4.3.1
blinker==1.7.0
# PySocks==1.7.1  # https://stackoverflow.com/questions/60254571/no-module-named-socks
selenium-wire==5.1.0
iso8601==2.1.0
typing-extensions==4.12.2
boto3==1.34.140
PyJWT==2.8.0
dictdiffer==0.9.0
StrEnum==0.4.15
